Combine server, state store and state machine
