+++ pygase/__init__.py
@@ -5,6 +5,6 @@ a complete set of high-level components for real-time networking for games.
 """
 
 from pygase.client import Client
-from pygase.server import Server
-from pygase.gamestate import GameState, GameStateStore, GameStateMachine
+from pygase.backend import Server, GameStateStore, GameStateMachine
+from pygase.gamestate import GameState
 from pygase.utils import get_available_ip_addresses
